refactor: extract edit bet service types to separate file and update GetEditableBets to use BetFilters
Move Movement/Transaction/EditBetService/EventBus/EditBetRequest/EditBetResponse structs from service.go to types.go. Replace GetByUserAndStatus call with GetByUserID using BetFilters in GetEditableBets method.
